Renamed main class. Added fitting and series acquisition tabs.
Series acquisition is currently being tested, working so far in limited cases.

Fitting works reliably, however the initial phase must be carefully specified to be 0 or pi for a sin or -sin waveform. This should be eliminated in future by implementing automatic finding of the initial gradient and removing the ability of a user specified phase.